Move tests to suitable files
